refactor(packages): move media host types next to their implementations
